Deal with non-strings in is_uuid?
In d971ad5ea1ae8945acad353a57b75561993cab2e the method was rewritten
with the intention of being compatible with Ruby 3.3 where =~ is only
defined on strings.

A problem with this is that /regex/.match?(nil) works, but
/regex/.match?(42) raises a TypeError. It would have been broken with
the old code on Ruby 3.3 as well, but only raised a deprecation warning
on prior versions.

Fixes: d971ad5ea1ae ("Fix warning on nil comparison")